+++ changelog.md
@@ -5,6 +5,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 ## v0.0.1 (unreleased)
 
 - [dev] Format JSON files
+- [build][dev] Upgrade Fork TS Checker Webpack Plugin
 - [dev] Add ESLint and stylelint configs and scripts
 - [dev] Suppress Webpack performance hints for the Storybook build
 - [dev] Enable test coverage thresholds